Add adaptive function bound computation (#17)
* Add adaptive function bound computation

This introduces a type-based approach that works slightly differently
than the current API. Upon construction of a `FDMethod` subtype, the
order of the method and the order of the derivative are checked for
agreement, and the grid and coefficients are computed and stored.
Instances of the types are callable, and doing so produces the value of
the derivative. It works recursively when adaptively computing the
bound, which results in a better step size and thus a (hopefully) more
robust estimate of the derivative.

* WIP: Add more things

* WIP: Store computational history in a History object

* Restructure to use the new API, add deprecations

This change consistutes a near rewrite of method handling.

* Add a max_step parameter to cap the step size

If it's too big, Bad Things can happen. See e.g. the added test.

* Pass max_step during adaptive steps as well