test/methods: (in progress) add explicit adapt parameter cases to Unitful tests
These seem to at least produce correct units, but the numerical value is
wrong for the adapt=0 case (not surprising, based on the documentation).

The question is, why the heck do things go wrong, when you do not give
an explicit adapt value, but leave it to the compiler-generated method
that inserts the value adapt=1 automatically.